Fix return types for SHFastIO.h (#32)

+++ src/SHFastIO.h
@@ -4,8 +4,8 @@
 
 class FastDigitalPin {
 private:
-	uint8_t bit = 0;
-	uint8_t port = 0;
+	uint32_t bit = 0;
+	uint32_t port = 0;
 	bool pinIsValid = false;
 public:
